# api/openapi.yaml
openapi: 3.1.0
info:
  title: AtApp Gateway API
  description: API Gateway interface for client-service interactions in AtApp college attendance system.
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /admin/users:
    post:
      summary: Create a user in the admin's own college
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
                - name
                - email
                - password
              properties:
                role:
                  type: string
                  enum: [student, teacher, admin]
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        '400':
          description: Missing fields or invalid role
        '403':
          description: Requires admin role
  /admin/departments:
    post:
      summary: Create a department in the admin's own college
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Department created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        '403':
          description: Requires admin role
  /admin/courses:
    post:
      summary: Create a course under a department
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - departmentId
                - name
                - code
              properties:
                departmentId:
                  type: string
                  format: uuid
                name:
                  type: string
                code:
                  type: string
                attendanceThresholdPct:
                  type: number
                  nullable: true
      responses:
        '201':
          description: Course created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        '403':
          description: Requires admin role
  /admin/sections:
    post:
      summary: Create a class section for a course
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - courseId
                - term
                - teacherId
              properties:
                courseId:
                  type: string
                  format: uuid
                term:
                  type: string
                teacherId:
                  type: string
                  format: uuid
                classroomBssid:
                  type: string
                  nullable: true
                classroomGeofenceLat:
                  type: number
                  nullable: true
                classroomGeofenceLng:
                  type: number
                  nullable: true
                classroomGeofenceRadiusM:
                  type: number
                  nullable: true
      responses:
        '201':
          description: Section created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        '403':
          description: Requires admin role
  /admin/import/users:
    post:
      summary: Bulk import users from a CSV file
      description: "CSV format: role,name,email,password"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Import summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                  message:
                    type: string
        '403':
          description: Requires admin role
  /admin/import/enrollments:
    post:
      summary: Bulk import student enrollments from a CSV file
      description: "CSV format: student_email,section_id"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Import summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                  message:
                    type: string
        '403':
          description: Requires admin role
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    PlatformAdminKey:
      type: apiKey
      in: header
      name: X-Platform-Admin-Key